F-11117 - Use size type for recipient depth

+++ src/wolfcose_hdr.c
@@ -645,7 +645,7 @@ int wolfCose_DecodeSkippedRecipient(WOLFCOSE_CBOR_CTX* ctx,
     int ret;
     size_t remaining = 1u;
     size_t stack[WOLFCOSE_CBOR_MAX_DEPTH];
-    unsigned int depth = 0u;
+    size_t depth = 0u;
     int firstRecipient = 1;
 
     if ((ctx == NULL) || (recipientAlg == NULL)) {
@@ -676,8 +676,8 @@ int wolfCose_DecodeSkippedRecipient(WOLFCOSE_CBOR_CTX* ctx,
                 ((nestedCount == 0u) || (nestedCount > ctx->bufSz))) {
                 ret = WOLFCOSE_E_CBOR_MALFORMED;
             }
-            if ((ret == WOLFCOSE_SUCCESS) && (depth >=
-                    (unsigned int)WOLFCOSE_CBOR_MAX_DEPTH)) {
+            if ((ret == WOLFCOSE_SUCCESS) &&
+                (depth >= (size_t)WOLFCOSE_CBOR_MAX_DEPTH)) {
                 ret = WOLFCOSE_E_CBOR_DEPTH;
             }
             if (ret == WOLFCOSE_SUCCESS) {
